Split api tests into more specific go files
