+++ api/api_test.go
@@ -1,22 +1,13 @@
 package api
 
 import (
-	"encoding/json"
-	"errors"
 	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-	"time"
-
 	"io"
 
 	"github.com/ONSdigital/dp-filter-api/api/datastoretest"
-	"github.com/ONSdigital/dp-filter-api/mocks"
 	"github.com/ONSdigital/dp-filter-api/models"
 	"github.com/ONSdigital/dp-filter-api/preview"
 	"github.com/ONSdigital/go-ns/common"
-	"github.com/gorilla/mux"
 	. "github.com/smartystreets/goconvey/convey"
 	"github.com/ONSdigital/dp-filter-api/filters"
 )
@@ -26,1497 +17,22 @@ const (
 	enablePrivateEndpoints = true
 	downloadServiceURL     = "http://localhost:23600"
 	downloadServiceToken   = "123wut"
-
 )
 
 var (
-	filterNotFoundResponse = filters.ErrFilterBlueprintNotFound.Error() + "\n"
+	filterNotFoundResponse       = filters.ErrFilterBlueprintNotFound.Error() + "\n"
 	filterOutputNotFoundResponse = filters.ErrFilterOutputNotFound.Error() + "\n"
-	dimensionNotFoundResponse = filters.ErrDimensionNotFound.Error() + "\n"
-	versionNotFoundResponse = filters.ErrVersionNotFound.Error() + "\n"
-	optionNotFoundResponse = filters.ErrOptionNotFound.Error() + "\n"
-	badRequestResponse = badRequest+"\n"
-	internalErrResponse = internalError+"\n"
+	dimensionNotFoundResponse    = filters.ErrDimensionNotFound.Error() + "\n"
+	versionNotFoundResponse      = filters.ErrVersionNotFound.Error() + "\n"
+	optionNotFoundResponse       = filters.ErrOptionNotFound.Error() + "\n"
+	badRequestResponse           = badRequest + "\n"
+	internalErrResponse          = internalError + "\n"
 )
 
 var previewMock = &datastoretest.PreviewDatasetMock{
 	GetPreviewFunc: func(filter *models.Filter, limit int) (*preview.FilterPreview, error) {
 		return &preview.FilterPreview{}, nil
 	},
-}
-
-func TestSuccessfulAddFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully create a filter blueprint", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} }`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully create a filter blueprint for an unpublished version", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"}, "dimensions":[{"name": "age", "options": ["27","33"]}]}`)
-		r := createAuthenticatedRequest("POST", "http://localhost:22100/filters", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully create a filter blueprint with dimensions", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"}, "dimensions":[{"name": "age", "options": ["27","33"]}]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	//	TODO check test doesn't actually write job to queue?
-	Convey("Successfully submit a filter blueprint", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} }`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters?submitted=true", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-}
-
-func TestFailedToAddFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} }`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When dataset API is unavailable, an internal error is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} }`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{InternalServerError: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When version does not exist, a not found error is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} }`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{VersionNotFound: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, versionNotFoundResponse)
-	})
-
-	Convey("When version is unpublished and the request is not authenticated, a bad request error is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"}, "dimensions":[{"name": "age", "options": ["27","33"]}]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When an invalid json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{")
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a empty json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{}")
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a json message is missing mandatory fields, a bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":"Census"}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a json message contains a dimension that does not exist, a bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} , "dimensions":[{"name": "weight", "options": ["27","33"]}]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimensions chosen: [weight]\n")
-	})
-
-	Convey("When a json message contains a dimension option that does not exist for a valid dimension, a bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"} , "dimensions":[{"name": "age", "options": ["29","33"]}]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimension options chosen: [29]\n")
-	})
-}
-
-func TestSuccessfulAddFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully create a dimension with an empty request body", t, func() {
-		reader := strings.NewReader("")
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully create a dimension with a request body but no options", t, func() {
-		reader := strings.NewReader("{}")
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully create a dimension with options", t, func() {
-		reader := strings.NewReader(`{"options":["27","33"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully create a dimension with options for an unpublished filter", t, func() {
-		reader := strings.NewReader(`{"options":["27","33"]}`)
-		r := createAuthenticatedRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-}
-
-func TestFailedToAddFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		reader := strings.NewReader(`{"options":["22","17"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When an invalid json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{")
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a filter blueprint does not exist, a not found is returned", t, func() {
-		reader := strings.NewReader(`{"options":["22","17"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When an unpublished filter blueprint does not exist, and the request is not authenticated, a not found is returned", t, func() {
-		reader := strings.NewReader(`{"options":["22","17"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When the dimension does not exist against the dataset filtered on, a bad request is returned", t, func() {
-		reader := strings.NewReader(`{"options":["22","17"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/wealth", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimensions chosen: [wealth]\n")
-	})
-
-	Convey("When a json body contains a dimension option that does not exist for a valid dimension, a bad request is returned", t, func() {
-		reader := strings.NewReader(`{"options":["22","33"]}`)
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimension options chosen: [22]\n")
-	})
-}
-
-func TestSuccessfulAddFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully add a dimension option to a filter", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/33", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-
-	Convey("Successfully add a dimension option to an unpublished filter", t, func() {
-		r := createAuthenticatedRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/33", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusCreated)
-	})
-}
-
-func TestFailedToAddFilterBlueprintDimensionOption_DimensionDoesNotExist(t *testing.T) {
-
-	Convey("When a dimension for filter blueprint does not exist, a bad request status is returned", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/notage/options/33", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InvalidDimensionOption: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimensions chosen: [notage]\n")
-	})
-}
-
-func TestFailedToAddFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/33", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When the filter blueprint does not exist, a bad request status is returned", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/33", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When the filter blueprint is unpublished, and the request is unauthenticated, a bad request status is returned", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/33", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When the dimension option for filter blueprint does not exist, a bad request status is returned", t, func() {
-		r, err := http.NewRequest("POST", "http://localhost:22100/filters/12345678/dimensions/age/options/66", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimension options chosen: [66]\n")
-	})
-}
-
-func TestSuccessfulGetFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a published filter blueprint with no authentication", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully get an unpublished filter blueprint with authentication", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filters/12345678", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToGetFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/1234568", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished, and the request is unauthenticated, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-}
-
-func TestSuccessfulUpdateFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully send a valid json message", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{ChangeInstanceRequest: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully send a valid json message with events and dataset version update", t, func() {
-
-		updateBlueprintData := `{"dataset":{"version":1}, "events":{"info":[{"time":"` + time.Now().String() +
-			`","type":"something changed","message":"something happened"}],"error":[{"time":"` + time.Now().String() +
-			`","type":"errored","message":"something errored"}]}}`
-
-		reader := strings.NewReader(updateBlueprintData)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{ChangeInstanceRequest: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully send a request to submit filter blueprint", t, func() {
-		reader := strings.NewReader("{}")
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312?submitted=true", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully send a request to submit an unpublished filter blueprint", t, func() {
-		reader := strings.NewReader("{}")
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filters/21312?submitted=true", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToUpdateFilterBlueprint(t *testing.T) {
-	t.Parallel()
-	Convey("When an invalid json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{")
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When an empty json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{}")
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a json message is sent to update filter blueprint that doesn't exist, a status of not found is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When no authentication is provided to update an unpublished filter, a not found is returned", t, func() {
-		reader := strings.NewReader(`{"dimensions":[]}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When a json message is sent to change the dataset version of a filter blueprint and the version does not exist, a status of bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":2}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{VersionNotFound: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, versionNotFoundResponse)
-	})
-
-	Convey("When a json message is sent to change the datset version of a filter blueprint and the current dimensions do not match, a status of bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":2}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimensions chosen: [time]\n")
-	})
-
-	Convey("When a json message is sent to change the dataset version of a filter blueprint and the current dimension options do not match, a status of bad request is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":2}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filters/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InvalidDimensionOption: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "incorrect dimension options chosen: [28]\n")
-	})
-}
-
-func TestSuccessfulGetFilterBlueprintDimensions(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a list of dimensions for a filter blueprint", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully get a list of dimensions for an unpublished filter blueprint", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filters/12345678/dimensions", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToGetFilterBlueprintDimensions(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/1234568/dimensions", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-}
-
-func TestSuccessfulGetFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a dimension for a filter blueprint, returns 204", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNoContent)
-	})
-
-	Convey("Successfully get a dimension for an unpublished filter blueprint, returns 204", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNoContent)
-	})
-}
-
-func TestFailedToGetFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/1234568/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a bad request is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished and request is unauthenticated, a bad request is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When dimension does not exist against filter blueprint, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{DimensionNotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, dimensionNotFoundResponse)
-	})
-}
-
-func TestSuccessfulGetFilterBlueprintDimensionOptions(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a list of dimension options for a filter blueprint", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/time/options", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully get a list of dimension options for an unpublished filter blueprint", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filters/12345678/dimensions/time/options", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToGetFilterBlueprintDimensionOptions(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/1234568/dimensions/1_age/options", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/123/dimensions/1_age/options", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished and the request is unauthenticated, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/123/dimensions/1_age/options", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When dimension does not exist against filter blueprint, a dimension not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age/options", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{DimensionNotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, dimensionNotFoundResponse)
-	})
-}
-
-func TestSuccessfulGetFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a single dimension option for a filter blueprint", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/time/options/2015", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNoContent)
-	})
-
-	Convey("Successfully get a single dimension option for an unpublished filter blueprint", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filters/12345678/dimensions/time/options/2015", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNoContent)
-	})
-}
-
-func TestFailedToGetFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/1234568/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a bad request is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished, a bad request is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When option does not exist against filter blueprint, an option not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filters/12345678/dimensions/age/options/notanage", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InvalidDimensionOption: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, optionNotFoundResponse)
-	})
-}
-
-func TestSuccessfulRemoveFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully remove a dimension for a filter blueprint, returns 200", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully remove a dimension for an unpublished filter blueprint, returns 200", t, func() {
-		r := createAuthenticatedRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToRemoveFilterBlueprintDimension(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/1234568/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a bad request is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished, and request is not authenticated, a bad request is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When dimension does not exist against filter blueprint, the response is idempotent and returns 200 OK", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{DimensionNotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestSuccessfulRemoveFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully remove a option for a filter blueprint, returns 200", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/time/options/2015", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully remove a option for an unpublished filter blueprint, returns 200", t, func() {
-		r := createAuthenticatedRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/time/options/2015", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToRemoveFilterBlueprintDimensionOption(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter blueprint does not exist, a bad request is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When filter blueprint is unpublished, and request is not authenticated, a bad request is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filterNotFoundResponse)
-	})
-
-	Convey("When dimension does not exist against filter blueprint, a not found is returned", t, func() {
-		r, err := http.NewRequest("DELETE", "http://localhost:22100/filters/12345678/dimensions/1_age/options/26", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{DimensionNotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, dimensionNotFoundResponse)
-	})
-}
-
-func TestSuccessfulGetFilterOutput(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully get a filter output from an unauthenticated request", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-
-		// Check private link is hidden for unauthenticated user
-		jsonResult := w.Body.Bytes()
-
-		filterOutput := &models.Filter{}
-		if err = json.Unmarshal(jsonResult, filterOutput); err != nil {
-			t.Logf("failed to marshal filte output json response, error: [%v]", err.Error())
-			t.Fail()
-		}
-
-		So(filterOutput.Downloads.CSV, ShouldResemble, &models.DownloadItem{HRef: "ons-test-site.gov.uk/87654321.csv", Private: "", Public: "csv-public-link", Size: "12mb"})
-		So(filterOutput.Downloads.XLS, ShouldResemble, &models.DownloadItem{HRef: "ons-test-site.gov.uk/87654321.xls", Private: "", Public: "xls-public-link", Size: "24mb"})
-	})
-
-	Convey("Successfully get a filter output from a request with an authorised download service token", t, func() {
-
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filter-outputs/12345678", nil)
-		r.Header.Add(common.DownloadServiceHeaderKey, downloadServiceToken)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-
-		// Check private link is NOT hidden from authenticated user
-		jsonResult := w.Body.Bytes()
-
-		filterOutput := &models.Filter{}
-		if err := json.Unmarshal(jsonResult, filterOutput); err != nil {
-			t.Logf("failed to marshal filte output json response, error: [%v]", err.Error())
-			t.Fail()
-		}
-
-		So(filterOutput.Downloads.CSV, ShouldResemble, &models.DownloadItem{HRef: "ons-test-site.gov.uk/87654321.csv", Private: "csv-private-link", Public: "csv-public-link", Size: "12mb"})
-		So(filterOutput.Downloads.XLS, ShouldResemble, &models.DownloadItem{HRef: "ons-test-site.gov.uk/87654321.xls", Private: "xls-private-link", Public: "xls-public-link", Size: "24mb"})
-	})
-
-	Convey("Successfully get an unpublished filter output", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filter-outputs/12345678", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToGetFilterOutput(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/1234568", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When filter output does not exist, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filters.ErrFilterOutputNotFound.Error()+"\n")
-	})
-
-	Convey("When filter output is unpublished and the request is unauthenticated, a not found is returned", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/12345678", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filters.ErrFilterOutputNotFound.Error()+"\n")
-	})
-}
-
-func TestSuccessfulUpdateFilterOutput(t *testing.T) {
-	t.Parallel()
-
-	Convey("Successfully update filter output when public csv download link is missing", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "public":"s3-public-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{MissingPublicLinks: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully update filter output when public xls download link is missing", t, func() {
-		reader := strings.NewReader(`{"downloads":{"xls":{"size":"12mb", "public":"s3-public-xls-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{MissingPublicLinks: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestSuccessfulUpdateFilterOutputUnpublished(t *testing.T) {
-	t.Parallel()
-
-	Convey("Successfully update filter output with private csv download link when version is unpublished", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "private": "s3-private-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-
-	Convey("Successfully update filter output with private xls download link when version is unpublished", t, func() {
-		reader := strings.NewReader(`{"downloads":{"xls":{"size":"12mb", "private":"s3-private-xls-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestFailedToUpdateFilterOutput(t *testing.T) {
-	t.Parallel()
-	Convey("When no data store is available, an internal error is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "public":"s3-public-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("When an invalid json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{")
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When an update to a filter output resource that does not exist, a not found is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "public":"s3-public-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-	})
-
-	Convey("When a empty json message is sent, a bad request is returned", t, func() {
-		reader := strings.NewReader("{}")
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, badRequestResponse)
-	})
-
-	Convey("When a json message contains fields that are not allowed to be updated, a forbidden status is returned", t, func() {
-		reader := strings.NewReader(`{"dataset":{"version":1, "edition":"1", "id":"1"}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusForbidden)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "Forbidden from updating the following fields: [dataset.id dataset.edition dataset.version]\n")
-	})
-
-	Convey("When a json message is sent to change a filter output with the wrong authorisation header, an unauthorised status is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "public":"s3-public-csv-location"}}}`)
-		r, err := http.NewRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusUnauthorized)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, errUnauthorised.Error()+"\n")
-	})
-
-	Convey("When a json message contains downloads object but current filter ouput has public csv download links already and version is published, than a forbidden status is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "public":"s3-public-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusForbidden)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "Forbidden from updating the following fields: [downloads.csv]\n")
-	})
-
-	Convey("When a json message contains downloads object but current filter ouput has public xls download links already and version is published, than a forbidden status is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"xls":{"href":"s3-xls-location","size":"12mb", "public":"s3-public-xls-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusForbidden)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "Forbidden from updating the following fields: [downloads.xls]\n")
-	})
-
-	Convey("When a json message contains private csv link but current filter ouput has private csv download links already and version is published, than a forbidden status is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"size":"12mb", "private":"s3-private-csv-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{MissingPublicLinks: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusForbidden)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "Forbidden from updating the following fields: [downloads.csv.private]\n")
-	})
-
-	Convey("When a json message contains private xls link but current filter ouput has private xls download links already and version is published, than a forbidden status is returned", t, func() {
-		reader := strings.NewReader(`{"downloads":{"xls":{"size":"12mb", "private":"s3-private-xls-location"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{MissingPublicLinks: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusForbidden)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "Forbidden from updating the following fields: [downloads.xls.private]\n")
-	})
-}
-
-func TestUpdateFilterOutput_PrivateEndpointsNotEnabled(t *testing.T) {
-
-	Convey("When private endpoints are not enabled, calling update on the filter output returns a 404 not found", t, func() {
-		reader := strings.NewReader(`{"downloads":{"csv":{"url":"s3-csv-location","size":"12mb"}}}`)
-		r := createAuthenticatedRequest("PUT", "http://localhost:22100/filter-outputs/21312", reader)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, false, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusMethodNotAllowed)
-	})
-}
-
-func TestSuccessfulGetPreview(t *testing.T) {
-	t.Parallel()
-	Convey("Successfully requesting a valid preview", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(previewMock.GetPreviewCalls()[0].Limit, ShouldEqual, 20)
-	})
-
-	Convey("Successfully requesting a valid preview for unpublished version filters", t, func() {
-		r := createAuthenticatedRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(previewMock.GetPreviewCalls()[0].Limit, ShouldEqual, 20)
-	})
-
-	Convey("Successfully requesting a valid preview with a new limit", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview?limit=10", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		previewMockForLimit := &datastoretest.PreviewDatasetMock{
-			GetPreviewFunc: func(filter *models.Filter, limit int) (*preview.FilterPreview, error) {
-				return &preview.FilterPreview{}, nil
-			},
-		}
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMockForLimit, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-		So(previewMockForLimit.GetPreviewCalls()[0].Limit, ShouldEqual, 10)
-	})
-}
-
-func TestFailedGetPreview(t *testing.T) {
-	t.Parallel()
-	Convey("Requesting a preview with invalid filter", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{NotFound: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, filters.ErrFilterOutputNotFound.Error()+"\n")
-	})
-
-	Convey("Requesting a preview with no mongodb database connection", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{InternalError: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, internalErrResponse)
-	})
-
-	Convey("Requesting a preview with no neo4j database connection", t, func() {
-		previewMockInternalError := &datastoretest.PreviewDatasetMock{
-			GetPreviewFunc: func(filter *models.Filter, limit int) (*preview.FilterPreview, error) {
-				return nil, errors.New("internal error")
-			},
-		}
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMockInternalError, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "internal server error\n")
-	})
-
-	Convey("Requesting a preview with no dimensions", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{BadRequest: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "no dimensions are present in the filter\n")
-	})
-
-	Convey("Requesting a preview with an invalid limit", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview?limit=a", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{}, &mocks.FilterJob{}, &mocks.DatasetAPI{}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "requested limit is not a number\n")
-	})
-
-	Convey("Requesting a preview with no authentication when the version is unpublished", t, func() {
-		r, err := http.NewRequest("GET", "http://localhost:22100/filter-outputs/21312/preview?limit=a", nil)
-		So(err, ShouldBeNil)
-
-		w := httptest.NewRecorder()
-		api := routes(host, mux.NewRouter(), &mocks.DataStore{Unpublished: true}, &mocks.FilterJob{}, &mocks.DatasetAPI{Unpublished: true}, previewMock, enablePrivateEndpoints, downloadServiceURL, downloadServiceToken)
-		api.router.ServeHTTP(w, r)
-		So(w.Code, ShouldEqual, http.StatusBadRequest)
-
-		response := w.Body.String()
-		So(response, ShouldResemble, "requested limit is not a number\n")
-	})
 }
 
 func createAuthenticatedRequest(method, url string, body io.Reader) *http.Request {
